Rename data_store to dataStore
